move websocket connection into api module, and move message subscription into corresponding feature module.
