--- app/utils/message-subscription.js
@@ -1,5 +1,0 @@
-/* @flow weak */
-export default function subscription() {
-  // console.log(this.subscribe);
-  console.log('subscribe');
-}
